+++ src/main/kotlin/io/arrowkt/arrowio/service/Common.kt
@@ -1,4 +1,4 @@
-package io.arrowkt.service
+package io.arrowkt.arrowio.service
 
 import arrow.data.EitherT
 import arrow.data.EitherTPartialOf
@@ -9,7 +9,7 @@ import arrow.data.extensions.kleisli.monad.monad
 import arrow.effects.ForIO
 import arrow.effects.IO
 import arrow.effects.extensions.io.monad.monad
-import io.arrowkt.repository.AccountRepository
+import io.arrowkt.arrowio.repository.AccountRepository
 
 val AccountOperationMonad = ReaderT.monad<EitherTPartialOf<ForIO, AccountServiceException>, AccountRepository>(
     EitherT.monad(IO.monad()))
