+++ src/main/kotlin/io/arrowkt/arrowio/service/InterestPostingService.kt
@@ -1,3 +1,3 @@
-package io.arrowkt.service
+package io.arrowkt.arrowio.service
 
 interface InterestPostingService<Account, Amount> : InterestCalculation<Account, Amount>, TaxCalculation<Amount>
