+++ src/main/kotlin/io/arrowkt/arrowio/service/ReportingService.kt
@@ -1,9 +1,9 @@
-package io.arrowkt.service
+package io.arrowkt.arrowio.service
 
 import arrow.data.EitherTPartialOf
 import arrow.data.Kleisli
 import arrow.effects.ForIO
-import io.arrowkt.repository.AccountRepository
+import io.arrowkt.arrowio.repository.AccountRepository
 
 typealias ReportOperation<A> = Kleisli<EitherTPartialOf<ForIO, AccountServiceException>, AccountRepository, A>
