+++ src/main/kotlin/io/arrowkt/arrowio/service/TaxCalculation.kt
@@ -1,4 +1,4 @@
-package io.arrowkt.service
+package io.arrowkt.arrowio.service
 
 import arrow.data.EitherTPartialOf
 import arrow.data.Kleisli
